requirements.txt
fastapi
uvicorn[standard]
httpx
pytest
pytest-asyncio
google-generativeai
pydantic-settings
python-dotenv
